Using tqdm, stop user translation, 2x resources
This replaces progress.Bar by tqdm, better performance and cleaner code.

We were mapping users id to a counter twice: once in the
simulate_task_allocation and another in the Task class itself. This made
us loose the original id -- not anymore.

Amount of total resources is now 2x the mean. This shoudn't matter as it
will be replaced by cli parameter soon...